+++ website_business_directory/doc/changelog.rst
@@ -1,3 +1,8 @@
+v1.0.1
+======
+* Featured listing feature
+* Bug fix related to new buiness form
+
 v1.0
 ====
 Initial Release